refactor to TokenGroup
