+++ src/app/components/TokenGroup/TokenGroup.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { TokenTypes } from '@/constants/TokenTypes';
 import { DeepKeyTokenMap, TokenTypeSchema } from '@/types/tokens';
-import { fireEvent, render, resetStore } from '../../../tests/config/setupTest';
-import { store } from '../store';
-import TokenTree from './TokenTree';
+import { fireEvent, render, resetStore } from '../../../../tests/config/setupTest';
+import { store } from '../../store';
+import TokenGroup from './TokenGroup';
 
 const displayType = 'GRID';
 const schema = {
@@ -63,13 +63,13 @@ const tokenValues = {
 const showNewForm = jest.fn();
 const showForm = jest.fn();
 
-describe('TokenTree', () => {
+describe('TokenGroup', () => {
   beforeEach(() => {
     resetStore();
   });
 
   it('should be able to collpase a token group', async () => {
-    const { getByTestId } = render(<TokenTree
+    const { getByTestId } = render(<TokenGroup
       displayType={displayType}
       tokenValues={tokenValues as DeepKeyTokenMap}
       showNewForm={showNewForm}
@@ -85,7 +85,7 @@ describe('TokenTree', () => {
   });
 
   it('should be able to expand a token group', async () => {
-    const { getByTestId } = render(<TokenTree
+    const { getByTestId } = render(<TokenGroup
       displayType={displayType}
       tokenValues={tokenValues as DeepKeyTokenMap}
       showNewForm={showNewForm}
@@ -102,7 +102,7 @@ describe('TokenTree', () => {
   });
 
   it('should be able to add a new token', async () => {
-    const result = render(<TokenTree
+    const result = render(<TokenGroup
       displayType={displayType}
       tokenValues={tokenValues as DeepKeyTokenMap}
       showNewForm={showNewForm}
