+++ src/app/components/TokenGroup/index.tsx
@@ -1,0 +1,1 @@
+export * as TokenGroup from './TokenGroup';
