+++ src/stitches.config.ts
@@ -140,6 +140,7 @@ export const stitchesInstance = createStitches({
       card: globalTokens.radii.lg,
       badge: globalTokens.radii.lg,
       full: globalTokens.radii.full,
+      tree: globalTokens.radii.lg,
     },
     shadows: {
       default: '0 1px 3px 0 rgba(0, 0, 0, 0.1), 0 1px 2px 0 rgba(0, 0, 0, 0.06)',
