+++ src/stories/TokenButton.stories.mdx
@@ -8,7 +8,6 @@ import { TokenTypes } from '@/constants/TokenTypes';
 import { BoxShadowTypes } from '@/constants/BoxShadowTypes';
 
 import { TokenButton } from '@/app/components/TokenButton';
-import TokenTree from '@/app/components/TokenTree';
 
 <Meta
   title="TokenButton"
